requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
anthropic>=0.25.0
openpyxl>=3.1.2
gitpython>=3.1.40
python-dotenv>=1.0.0
pandas>=2.2.0
plotly>=5.20.0
tqdm>=4.66.0
chardet>=5.2.0
bandit>=1.7.8
pyflakes>=3.2.0
requests